Enhanced FlexStore demo to use new logTarget property for l10nInject logging.
The FlexStore LocalizationMap now assigns a logTarget using the new StaticClassFactory and a log level of WARN. This works with the updated l10nInjection code to provide robust logging from within the l10n  engine.

Modified Registration, FlexStore, and Intranet demos projects to now use the l10nInjection library project instead of the l10nInjection.swc in the libs directory.